requirements.txt
flask==3.1.3
sentence-transformers==5.4.1
faiss-cpu==1.13.2
pandas==3.0.2
numpy==2.4.4
gunicorn==25.3.0
